pyproject.toml
[tool.poetry]
name = "wallet"
version = "0.1.0"
description = "Wallet API"
authors = ["Maria <maria.murmantseva@gmail.com>"]
readme = "README.md"
packages = [{ include = "wallet" }]

[tool.poetry.scripts]
prepare = "wallet.cli:prepare"
service = "wallet.cli:service"
token = "wallet.cli:token"

[tool.poetry.dependencies]
python = "^3.11"

fastapi = "^0.115.6"
uvicorn = "^0.34.0"

pydantic = "^2.10.4"
pydantic-settings = "^2.7.1"

sqlmodel = "^0.0.22"
asyncpg = "^0.30.0"

httpx = "^0.28.1"
pyjwt = {extras = ["crypto"], version = "^2.10.1"}
click = "^8.1.8"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
ruff = "^0.8.4"
mypy = "^1.14.1"
pytest-asyncio = "^0.25.1"
pytest-httpx = "^0.35.0"
python-on-whales = "^0.74.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"

[tool.ruff]
target-version = "py311"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "COM812", "D105", "D107", "D203", "D211", "D212", "EM", "FBT002", "ISC001",
    "TC002", "TRY003"
]

[tool.ruff.lint.per-file-ignores]
"tests/**.py" = ["ARG001", "D", "S101"]

[tool.roof.format]
ignore = ["COM812", "D203", "ISC001"]

[tool.mypy]
strict = true
plugins = "pydantic.mypy"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
